+++ CHANGELOG.rst
@@ -20,6 +20,7 @@ Changed
 -------
 * Improve NN metrics printing.
 * Feature reduction in TrendLengthData.
+* Save NNState as EmbeddedDocument.
 
 ==================
 1.1.0 - 2021-01-21